Created a new package named model to separate the model classes and the dao implementations
