+++ core_java/stockquote/stockquote/src/main/java/ca/jrvs/apps/stockquote/dao/PositionDao.java
@@ -1,6 +1,7 @@
 package ca.jrvs.apps.stockquote.dao;
 
 
+import ca.jrvs.apps.stockquote.model.Position;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
