+++ core_java/stockquote/stockquote/src/main/java/ca/jrvs/apps/stockquote/dao/QuoteDao.java
@@ -1,5 +1,6 @@
 package ca.jrvs.apps.stockquote.dao;
 
+import ca.jrvs.apps.stockquote.model.Quote;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
