+++ core_java/stockquote/stockquote/src/main/java/ca/jrvs/apps/stockquote/dao/QuoteHttpHelper.java
@@ -1,5 +1,6 @@
 package ca.jrvs.apps.stockquote.dao;
 
+import ca.jrvs.apps.stockquote.model.Quote;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
