+++ core_java/stockquote/stockquote/src/main/java/ca/jrvs/apps/stockquote/model/Position.java
@@ -1,4 +1,4 @@
-package ca.jrvs.apps.stockquote.dao;
+package ca.jrvs.apps.stockquote.model;
 
 public class Position {
