+++ core_java/stockquote/stockquote/src/test/java/QuoteHttpHelperTest.java
@@ -2,7 +2,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import ca.jrvs.apps.stockquote.dao.Quote;
+import ca.jrvs.apps.stockquote.model.Quote;
 import ca.jrvs.apps.stockquote.dao.QuoteHttpHelper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
